+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changelog
 * Add new command :command:`ngs-bam2fq`.
 * Add new command :command:`fa-filter`.
 * Update :meth:`pycov.CovFrame.plot_region` method to raise an error if the CovFrame is empty.
+* Update :meth:`pyvcf.VcfFrame.filter_*` methods so that they don't raise an error when the VcfFrame is empty.
 
 0.27.0 (2021-11-20)
 -------------------